chore: show which file has hit an error
string value owned by `path` lives only as long as this particular iteration of the `for` loop before its replaced by the next `path` argument - tidying up by Rust. This means a refeence to `path` will only be valid until the next loop iteration. `context` howevr takes a variable bounded by `'static` lifetime ie a value guaranteed to be around for the lifetime of the program and can potentially be retuned from this function. This means we need to make a copy of the string value and give this to `context` so that the error can refer to it.